Add rules so that applications behave better
QjackCtl doesn't automatically float so I add a rule to make it float by
default. I also make it so it spawns on the tag "1". I make mpv spawn in
all tags in floating since it doesn't make sense to have it tiled. One
shouldn't be multitasking while watching media. I add it in all tags so
I don't forget something playing in the background. Also mpv seems to
get synchronization problems if the video plays on a tag that isn't
visible.